Move card/ to common/
