+++ components/carte-etat/carte-etat-component.tsx
@@ -19,7 +19,7 @@ import { connect, ConnectedProps } from "react-redux";
 import { fetchSimPop, simulateCasTypes } from "../../redux/actions";
 // eslint-disable-next-line no-unused-vars
 import { RootState } from "../../redux/reducers";
-import { Card } from "../card";
+import { Card } from "../common";
 import BarChart from "./bar-chart";
 import styles2 from "./carte-etat-component.module.scss";
 import SimpopTableurInfosDeciles from "./simpop-tableur-infos-deciles";
