+++ components/cartes-impact/simple-card/simple-card-component.jsx
@@ -16,7 +16,7 @@ import { get } from "lodash";
 import PropTypes from "prop-types";
 import React, { Fragment } from "react";
 
-import { Card } from "../../card";
+import { Card } from "../../common";
 import DoublePalmTreeIcon from "../../icons/double-palm-tree";
 import { NeutralTooltip } from "../../tooltips";
 import { formatNumber } from "../../utils";
