+++ components/dotations/results/commune-summary/CommuneSummary.tsx
@@ -11,7 +11,7 @@ import { connect, ConnectedProps } from "react-redux";
 // eslint-disable-next-line no-unused-vars
 import { RootState } from "../../../../redux/reducers";
 import { Parameter, ResultValues } from "../../../articles-inputs/parameter";
-import { Card } from "../../../card";
+import { Card } from "../../../common";
 import styles from "./CommuneSummary.module.scss";
 
 const mapStateToProps = ({ results }: RootState) => ({
